Feat: 채팅방 커버 페이지 로딩 및 에러 핸들링
